% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bakers_percentage.R
\name{calculate_bakers_percentage}
\alias{calculate_bakers_percentage}
\title{Calculate Baker's Percentage}
\usage{
calculate_bakers_percentage(.df)
}
\arguments{
\item{.df}{A data frame containing the ingredients, their unit, and the
weight}
}
\value{
A data frame with the ingredients, their weight, and the baker's
percentage
}
\description{
Calculates the baker's percentage of ingredients in a given data frame for a
recipe.
}
\examples{
recipe <- data.frame(ingredient = c("flour", "water", "salt"),
                 unit = c("grams", "grams", "grams"),
                 weight = c(100, 60, 2))

calculate_bakers_percentage(recipe)
}
