% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourdough.R
\name{country_sourdough}
\alias{country_sourdough}
\alias{sourdough_baguette_dough}
\title{Tartine Sourdough}
\usage{
country_sourdough(loaves = 2, base_loaves = 2)

sourdough_baguette_dough(baguettes = 8, base_baguettes = 8)
}
\arguments{
\item{loaves}{Number of loaves of Tartine Country Sourdough}

\item{base_loaves}{Base number of loaves for Tartine Country Sourdough}

\item{baguettes}{Number of baguettes of Tartine Sourdough Baguettes}

\item{base_baguettes}{Base number of baguettes for Tartine Sourdough
Baguettes}
}
\value{
A tibble containing the standardized ingredient list
}
\description{
Generate ingredient lists for Tartine Country Sourdough and Tartine Sourdough
Baguettes.
}
\examples{
country_sourdough()
country_sourdough(loaves = 6, base_loaves = 3)
sourdough_baguette_dough(baguettes = 4)
}
