% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pizza.R
\name{new_york_pizza_dough}
\alias{new_york_pizza_dough}
\title{Create a recipe for New York-style pizza dough}
\usage{
new_york_pizza_dough(pizzas = 2)
}
\arguments{
\item{pizzas}{The number of pizzas to make. Default is 2.}
}
\value{
A tibble with ingredients, weights, and units.
}
\description{
This function creates a recipe for New York-style pizza dough.
}
\examples{
new_york_pizza_dough()
new_york_pizza_dough(pizzas = 4)
}
